Moving run_cmd from build.py to tools/cmds.py
Makes it easier to be reused.
